requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
mysqlclient==2.0.3
Pillow==8.3.1
pycodestyle==2.7.0
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
